product_importer/requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
django-cors-headers
django-extensions
python-dotenv
requests
